Rename file and component Offer -> QuoteData (in order for it to NOT be confused with the actual offer page)
